Began putting systems in place for better generation
Refactored a lot of platform generator to handle spawning rooms along
with sprinkling platforms in between. Still a lot of work needs to be
done.

Generator should work the same as before for now except for cleaning up
platforms as that function is currently being replaced and the new one
hasn't been implemented yet.