Sound Effect Update (Dialogue Text Sound)
This is a very, very short blip for text. To my understanding, you
wanted to play it for each character as it came up. I don't know for
sure if this is what you were looking for or what you meant, so let me
know if you wanted something slightly different.